refactor: move dupe file
