+++ src/workers/crypto/crypto.ts
@@ -13,14 +13,47 @@ import { Event, nip44 } from 'nostr-tools';
 import { toWithError } from '@/utils/async';
 import {
   bufToHex,
+  generateNewMnemonic,
   hexToUint8Array,
   restoreMnemonic,
   secp256k1GenerateProtectedKeyPair,
 } from '@/utils/crypto';
 import { decryptMessage, encryptMessage } from '@/utils/nostr';
 
-import { createNewWallet, generateLiquidDescriptor } from '../crypto';
-import { CryptoWorkerMessage, CryptoWorkerResponse } from './types';
+import {
+  CryptoNewWalletPayload,
+  CryptoWorkerMessage,
+  CryptoWorkerResponse,
+} from './types';
+
+export const createNewWallet = async (
+  masterKey: string
+): Promise<CryptoNewWalletPayload> => {
+  const { mnemonic, protectedMnemonic } = generateNewMnemonic(masterKey);
+
+  const { publicKey, protectedPrivateKey } =
+    secp256k1GenerateProtectedKeyPair(masterKey);
+
+  const liquidDescriptor = await generateLiquidDescriptor(mnemonic);
+
+  return {
+    protectedMnemonic: protectedMnemonic,
+    liquidDescriptor,
+    secp256k1_key_pair: {
+      public_key: publicKey,
+      protected_private_key: protectedPrivateKey,
+    },
+  };
+};
+
+const generateLiquidDescriptor = async (mnemonic: string) => {
+  const network = Network.mainnet();
+
+  const signer = new Signer(new Mnemonic(mnemonic), network);
+  const wolletDescriptor = signer.wpkhSlip77Descriptor().toString();
+
+  return wolletDescriptor;
+};
 
 const signPset = (mnemonic: string, descriptor: string, pset: string) => {
   const network = Network.mainnet();
